control: Move
