+++ pkg/control/pod_control.go
@@ -17,7 +17,7 @@ limitations under the License.
 // Most code here is copied from Kubernetes, while we change the
 // logic about name generation to generate name with type and index.
 
-package controller
+package control
 
 import (
 	"fmt"
