+++ pkg/control/service_ref_manager_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package controller
+package control
 
 import (
 	"reflect"
@@ -22,6 +22,7 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
 
+	tfv1alpha2 "github.com/kubeflow/tf-operator/pkg/apis/tensorflow/v1alpha2"
 	"github.com/kubeflow/tf-operator/pkg/generator"
 	"github.com/kubeflow/tf-operator/pkg/util/testutil"
 )
@@ -54,7 +55,7 @@ func TestClaimServices(t *testing.T) {
 				manager: NewServiceControllerRefManager(&FakeServiceControl{},
 					tfJob,
 					tfJobLabelSelector,
-					controllerKind,
+					tfv1alpha2.SchemeGroupVersionKind,
 					func() error { return nil }),
 				services: []*v1.Service{testutil.NewBaseService("service1", tfJob, t), testService},
 				claimed:  []*v1.Service{testutil.NewBaseService("service1", tfJob, t)},
@@ -80,7 +81,7 @@ func TestClaimServices(t *testing.T) {
 				manager: NewServiceControllerRefManager(&FakeServiceControl{},
 					controller,
 					controllerLabelSelector,
-					controllerKind,
+					tfv1alpha2.SchemeGroupVersionKind,
 					func() error { return nil }),
 				services: []*v1.Service{testService1, testService2},
 				claimed:  nil,
@@ -104,7 +105,7 @@ func TestClaimServices(t *testing.T) {
 				manager: NewServiceControllerRefManager(&FakeServiceControl{},
 					controller,
 					controllerLabelSelector,
-					controllerKind,
+					tfv1alpha2.SchemeGroupVersionKind,
 					func() error { return nil }),
 				services: []*v1.Service{testutil.NewBaseService("service1", controller, t), testService2},
 				claimed:  []*v1.Service{testutil.NewBaseService("service1", controller, t)},
@@ -126,7 +127,7 @@ func TestClaimServices(t *testing.T) {
 				manager: NewServiceControllerRefManager(&FakeServiceControl{},
 					controller,
 					controllerLabelSelector,
-					controllerKind,
+					tfv1alpha2.SchemeGroupVersionKind,
 					func() error { return nil }),
 				services: []*v1.Service{testutil.NewBaseService("service1", controller, t), testutil.NewBaseService("service2", controller2, t)},
 				claimed:  []*v1.Service{testutil.NewBaseService("service1", controller, t)},
@@ -148,7 +149,7 @@ func TestClaimServices(t *testing.T) {
 				manager: NewServiceControllerRefManager(&FakeServiceControl{},
 					controller,
 					controllerLabelSelector,
-					controllerKind,
+					tfv1alpha2.SchemeGroupVersionKind,
 					func() error { return nil }),
 				services: []*v1.Service{testutil.NewBaseService("service1", controller, t), testService2},
 				claimed:  []*v1.Service{testutil.NewBaseService("service1", controller, t)},
@@ -175,7 +176,7 @@ func TestClaimServices(t *testing.T) {
 				manager: NewServiceControllerRefManager(&FakeServiceControl{},
 					controller,
 					controllerLabelSelector,
-					controllerKind,
+					tfv1alpha2.SchemeGroupVersionKind,
 					func() error { return nil }),
 				services: []*v1.Service{testService1, testService2},
 				claimed:  []*v1.Service{testService1},
