+++ pkg/controller.v2/controller.go
@@ -40,6 +40,7 @@ import (
 	tfjobinformers "github.com/kubeflow/tf-operator/pkg/client/informers/externalversions"
 	tfjobinformersv1alpha2 "github.com/kubeflow/tf-operator/pkg/client/informers/externalversions/kubeflow/v1alpha2"
 	tfjoblisters "github.com/kubeflow/tf-operator/pkg/client/listers/kubeflow/v1alpha2"
+	"github.com/kubeflow/tf-operator/pkg/control"
 )
 
 const (
@@ -86,7 +87,7 @@ type TFJobController struct {
 	podControl controller.PodControlInterface
 
 	// serviceControl is used to add or delete services.
-	serviceControl ServiceControlInterface
+	serviceControl control.ServiceControlInterface
 
 	// kubeClientSet is a standard kubernetes clientset.
 	kubeClientSet kubeclientset.Interface
@@ -171,12 +172,12 @@ func NewTFJobController(
 	eventBroadcaster.StartRecordingToSink(&typedcorev1.EventSinkImpl{Interface: kubeClientSet.CoreV1().Events("")})
 	recorder := eventBroadcaster.NewRecorder(scheme.Scheme, v1.EventSource{Component: controllerName})
 
-	realPodControl := RealPodControl{
+	realPodControl := control.RealPodControl{
 		KubeClient: kubeClientSet,
 		Recorder:   eventBroadcaster.NewRecorder(scheme.Scheme, v1.EventSource{Component: controllerName}),
 	}
 
-	realServiceControl := RealServiceControl{
+	realServiceControl := control.RealServiceControl{
 		KubeClient: kubeClientSet,
 		Recorder:   eventBroadcaster.NewRecorder(scheme.Scheme, v1.EventSource{Component: controllerName}),
 	}
